Cycles: Stop using OpenImageIO texture cache for OSL
This will be fully replaced by our own texture cache in the future. Basic tests should
pass but there is missing functionality like environment() and gettextureinfo(), while
some more advanced OIIO features are not expected to be supported anymore.

UDIMs are currently broken, the kernel implementation of these will need to be moved
from SVM to the image texture code.